organize source packages
