+++ android/src/main/java/com/klarna/mobile/sdk/reactnative/KlarnaPaymentViewPackage.java
@@ -8,6 +8,7 @@ import com.facebook.react.ReactPackage;
 import com.facebook.react.bridge.NativeModule;
 import com.facebook.react.bridge.ReactApplicationContext;
 import com.facebook.react.uimanager.ViewManager;
+import com.klarna.mobile.sdk.reactnative.payments.KlarnaPaymentViewManager;
 
 import java.util.Collections;
 import java.util.List;
