+++ android/src/main/java/com/klarna/mobile/sdk/reactnative/payments/KlarnaPaymentEvent.java
@@ -1,4 +1,4 @@
-package com.klarna.mobile.sdk.reactnative;
+package com.klarna.mobile.sdk.reactnative.payments;
 
 import androidx.annotation.IdRes;
 import androidx.annotation.NonNull;
@@ -7,10 +7,6 @@ import androidx.annotation.Nullable;
 import com.facebook.react.bridge.Arguments;
 import com.facebook.react.bridge.WritableMap;
 import com.facebook.react.uimanager.events.Event;
-import com.facebook.react.uimanager.events.RCTEventEmitter;
-
-import java.util.HashMap;
-import java.util.Map;
 
 /**
  * A `KlarnaPaymentEvent` builds the event that will be eventually sent via `on<event>`.
